requirements.txt
web3==6.0.0
py-solc-x==1.1.1
flask==2.0.1
python-dotenv==0.19.0
textblob==0.15.3
werkzeug==2.0.1
ipfshttpclient==0.8.0
requests==2.26.0